feat: Places curated matches in a collapsible element
This makes it possible to hide the curated matches and view the computed
matches, if the curated matches are not what you are looking for.
