+++ src/components/CuratedResults.jsx
@@ -1,9 +1,5 @@
-import { useState, useEffect } from "react";
 import PropTypes from "prop-types";
 import { Card, Table, Typography } from "antd";
-import { Auth, API } from "aws-amplify";
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { faBookmark } from "@fortawesome/pro-solid-svg-icons";
 import { Link } from "react-router-dom";
 
 import HelpButton from "./Help/HelpButton";
@@ -35,89 +31,7 @@ const columns = [
   },
 ];
 
-function filterAndSortCuratedMatches(matches) {
-  // expand the matches.
-  const expanded = matches.flatMap((match) => {
-    if (match.itemType === "line_name") {
-      return match.matches.map((m) => ({
-        name: match.name,
-        confidence: m.annotation,
-        anatomicalRegion: m.region,
-        cellType: m.cell_type,
-      }));
-    }
-    if (match.itemType === "cell_type") {
-      return match.matches.map((m) => ({
-        name: m.line,
-        confidence: m.annotation,
-        anatomicalRegion: m.region,
-        cellType: match.name,
-      }));
-    }
-    return [];
-  });
-
-  // strip duplicates if cellType and name are the same
-  const deduped = expanded.filter(
-    (v, i, a) =>
-      a.findIndex((t) => t.cellType === v.cellType && t.name === v.name) === i,
-  );
-  // sort by name and then confidence, where confident comes before candidate.
-  deduped.sort((a, b) => {
-    if (a.confidence === "Confident" && b.confidence === "Candidate") {
-      return -1;
-    }
-    if (a.confidence === "Candidate" && b.confidence === "Confident") {
-      return 1;
-    }
-
-    if (a.name < b.name) {
-      return -1;
-    }
-    if (a.name > b.name) {
-      return 1;
-    }
-    return 0;
-  });
-  return deduped;
-}
-
-export default function CuratedResults({ searchTerm }) {
-  const [results, setResults] = useState([]);
-  const [loadError, setLoadError] = useState(false);
-
-  // use the searchTerm to fetch curated results from DynamoDB?
-  useEffect(() => {
-    if (searchTerm !== "") {
-      Auth.currentCredentials().then(() => {
-        setLoadError(false);
-        API.get("SearchAPI", "/curated_matches", {
-          queryStringParameters: {
-            q: searchTerm,
-          },
-        })
-          .then((response) => {
-            if (response.matches.length > 0) {
-              const sorted = filterAndSortCuratedMatches(response.matches);
-              setResults(sorted);
-            }
-          })
-          .catch((error) => {
-            setLoadError(error);
-          });
-      });
-    }
-  }, [searchTerm]);
-
-  // if we found anything, then display it in a table above the rest of the results.
-  // if we're still waiting for the fetch to complete, then display a loading spinner.
-  // if there was an error, then display an error message.
-
-  // if the user has not yet entered a search term, then don't display anything.
-  if (!searchTerm) {
-    return null;
-  }
-
+export default function CuratedResults({ results, loadError }) {
   // if we didn't find anything, then don't display anything.
   if (results.length === 0) {
     return null;
@@ -145,27 +59,10 @@ export default function CuratedResults({ searchTerm }) {
     );
   }
 
-  return (
-    <Card
-      title="Curated Matches"
-      extra={<HelpButton target="CuratedResults" />}
-      style={{ marginBottom: "2em", position: "relative" }}
-    >
-      <FontAwesomeIcon
-        icon={faBookmark}
-        style={{
-          position: "absolute",
-          top: "-3px",
-          left: "5px",
-          color: "#f00",
-          fontSize: "1.5em",
-        }}
-      />
-      <Table columns={columns} dataSource={results} pagination={false} />
-    </Card>
-  );
+  return <Table columns={columns} dataSource={results} pagination={false} />;
 }
 
 CuratedResults.propTypes = {
-  searchTerm: PropTypes.string.isRequired,
+  results: PropTypes.array.isRequired,
+  loadError: PropTypes.bool.isRequired,
 };
